chore: Fix cache test and clean up warnings
