+++ packages/cache-tools/src/util/types.ts
@@ -6,4 +6,4 @@ export type Optional<T> = T | undefined
 /**
  * Simple async function type
  */
-export type AsyncFunction<R> = (...args: any[]) => Promise<R>
+export type AsyncFunction<R> = (...args: never[]) => Promise<R>
